+++ src/main/java/com/DTOLayer/DTOEntity/orderDTO/OrderDTOItemCount.java
@@ -1,12 +1,14 @@
 package com.DTOLayer.DTOEntity.orderDTO;
 
 import com.DTOLayer.DTOEntity.ItemDTO;
+import com.DTOLayer.DTOEntity.RestaurantDTO;
+import com.DTOLayer.DTOEntity.restaurantDTO.RestaurantDTOforOrder;
 import com.model.Entity.Item;
 
 class OrderDTOItemCount {
     private ItemDTO item;
     private int count;
-
+    private RestaurantDTOforOrder restaurant;
     public int getCount() {
         return count;
     }
@@ -26,5 +28,14 @@ class OrderDTOItemCount {
     public OrderDTOItemCount(Item item){
         this.item = new ItemDTO(item);
         count = 1;
+        this.restaurant = new RestaurantDTOforOrder(item.getRestaurant());
+    }
+
+    public RestaurantDTOforOrder getRestaurant() {
+        return restaurant;
+    }
+
+    public void setRestaurant(RestaurantDTOforOrder restaurant) {
+        this.restaurant = restaurant;
     }
 }
