--- src/main/java/com/DTOLayer/JSON/OrderDTO.java
@@ -1,4 +1,0 @@
-package com.DTOLayer.JSON;
-
-public class OrderDTO {
-}
