--- src/main/java/com/DTOLayer/JSON/Restaurant.java
@@ -1,8 +1,0 @@
-package com.DTOLayer.JSON;
-
-public class Restaurant {
-    int id;
-    String name;
-    String description;
-
-}
