--- src/main/java/com/DTOLayer/JSON/RestaurantDTO.java
@@ -1,7 +1,0 @@
-package com.DTOLayer.JSON;
-
-/**
- * Created by employee on 4/15/16.
- */
-public class RestaurantDTO {
-}
